+++ lithops/serverless/backends/code_engine/code_engine.py
@@ -61,8 +61,14 @@ class CodeEngineBackend:
         contexts = config.list_kube_config_contexts(config_file=self.kubecfg)
         current_context = contexts[1].get('context')
         self.namespace = current_context.get('namespace', 'default')
+        logger.debug("Set namespace to {}".format(self.namespace))
         self.cluster = current_context.get('cluster')
-        self.region = self.cluster.split('//')[1].split('.')[1]
+        logger.debug("Set cluster to {}".format(self.cluster))
+
+        try:
+            self.region = self.cluster.split('//')[1].split('.')[1]
+        except Exception:
+            self.region = ''
 
         self.job_def_ids = set()